Revert part of original fix for 6841036
The original fix forces all the function decls to
be marked as not wanting IR generation if flag_use_rtl_backend
is set to -1. This information get written out to PCH.
Later when we process the PCH this may not be what we want.
Do not force setting the bit on the function decl.