config: Do not force runpath entries unless necessary
Sun patched their GCC3 to force /lib:/usr/lib:/usr/sfw/lib into the runpath, in
that order, such that /usr/sfw libraries did not take precedence over system
libraries, but nevertheless libgcc did not need a specific -R entry to be added
to the command line when building.

This has caused us nothing but trouble when patching it forward, since suddenly
we need our _own_ prefix in thre too, and it must beat /usr/sfw, and because
it's using -R it means that LD_RUN_PATH doesn't function, and because it's
unexpected it causes confusion for 3rd party software.

This change only uses -R to force the runpath in the BSD compat environment, as
GCC upstream does, leaving the sole artifact of Sun's change being the addition
of /lib to the default library search path. (-YP)