msave-args: preserve integer function arguments passed in registers
Similar to on amd64, it would be convenient if we could preserve call-time
function arguments in the stack frame for use by the debuggers.

Implement a prototype of that which works by telling the compiler that the
argument registers (x0-x7) are callee saved, and must be preserved, which does
what we want at, though is perhaps not the most future-proof.